+++ cognite/experimental/_api/geospatial.py
@@ -4,6 +4,8 @@ import numbers
 from typing import Any, Dict, Generator, List, Union
 
 from cognite.client._api_client import APIClient
+from cognite.client.exceptions import CogniteConnectionError
+from requests.exceptions import ChunkedEncodingError
 
 from cognite.experimental.data_classes.geospatial import (
     CoordinateReferenceSystem,
@@ -68,8 +70,8 @@ class ExperimentalGeospatialAPI(APIClient):
                 >>> from cognite.experimental.data_classes.geospatial import FeatureType
                 >>> c = CogniteClient()
                 >>> feature_types = [
-                ...     FeatureType(external_id="wells", attributes={"location": {"type": "POINT", "srid": 4326}})
-                ...     FeatureType(external_id="pipelines", attributes={"location": {"type": "LINESTRING", "srid": 2001}})
+                ...     FeatureType(external_id="wells", properties={"location": {"type": "POINT", "srid": 4326}})
+                ...     FeatureType(external_id="pipelines", properties={"location": {"type": "LINESTRING", "srid": 2001}})
                 ... ]
                 >>> res = c.geospatial.create_feature_types(feature_types)
         """
@@ -78,13 +80,13 @@ class ExperimentalGeospatialAPI(APIClient):
         )
 
     @_with_cognite_domain
-    def delete_feature_types(self, external_id: Union[str, List[str]], force: bool = False) -> None:
+    def delete_feature_types(self, external_id: Union[str, List[str]], recursive: bool = False) -> None:
         """`Delete one or more feature type`
         <https://pr-1323.specs.preview.cogniteapp.com/v1.json.html#operation/deleteFeatureTypes>
 
         Args:
             external_id (Union[str, List[str]]): External ID or list of external ids
-            force (bool): if `true` the features will also be dropped
+            recursive (bool): if `true` the features will also be dropped
 
         Returns:
             None
@@ -97,9 +99,12 @@ class ExperimentalGeospatialAPI(APIClient):
                 >>> c = CogniteClient()
                 >>> c.geospatial.delete_feature_types(external_id=["wells", "pipelines"])
         """
-        params = {"force": "true"} if force else None
+        extra_body_fields = {"recursive": True} if recursive else {}
         return self._delete_multiple(
-            external_ids=external_id, wrap_ids=True, resource_path=f"{self._RESOURCE_PATH}/featuretypes", params=params
+            external_ids=external_id,
+            wrap_ids=True,
+            resource_path=f"{self._RESOURCE_PATH}/featuretypes",
+            extra_body_fields=extra_body_fields,
         )
 
     @_with_cognite_domain
@@ -160,17 +165,17 @@ class ExperimentalGeospatialAPI(APIClient):
 
         Examples:
 
-            Add one attribute to a feature type:
+            Add one property to a feature type:
 
                 >>> from cognite.experimental import CogniteClient
                 >>> c = CogniteClient()
-                >>> res = c.geospatial.update_feature_types(update=FeatureTypeUpdate(external_id="wells", add=AttributeAndSearchSpec(attributes={"altitude": {"type": "DOUBLE"}})))
+                >>> res = c.geospatial.update_feature_types(update=FeatureTypeUpdate(external_id="wells", add=PropertyAndSearchSpec(properties={"altitude": {"type": "DOUBLE"}})))
         """
         if isinstance(update, FeatureTypeUpdate):
             update = [update]
 
         mapper = lambda it: {
-            "attributes": None if not hasattr(it, "attributes") else {"add": it.attributes},
+            "properties": None if not hasattr(it, "properties") else {"add": it.properties},
             "searchSpec": None if not hasattr(it, "search_spec") else {"add": it.search_spec},
         }
         json = {"items": [{"externalId": it.external_id, "update": mapper(it.add)} for it in update]}
@@ -206,8 +211,10 @@ class ExperimentalGeospatialAPI(APIClient):
                 >>> res = c.geospatial.create_features(my_feature_types, Feature(external_id="my_feature", temperature=12.4))
         """
         resource_path = self._feature_resource_path(feature_type)
-        params = {"allowCrsTransformation": "true"} if allow_crs_transformation else None
-        return self._create_multiple(items=feature, resource_path=resource_path, cls=FeatureList, params=params)
+        extra_body_fields = {"allowCrsTransformation": "true"} if allow_crs_transformation else {}
+        return self._create_multiple(
+            items=feature, resource_path=resource_path, cls=FeatureList, extra_body_fields=extra_body_fields
+        )
 
     @_with_cognite_domain
     def delete_features(self, feature_type: FeatureType, external_id: Union[str, List[str]] = None) -> None:
@@ -291,15 +298,17 @@ class ExperimentalGeospatialAPI(APIClient):
         # updates for feature are not following the patch structure from other resources
         # they are more like a replace so an update looks like a feature creation (yeah, borderline ?)
         resource_path = self._feature_resource_path(feature_type) + "/update"
-        params = {"allowCrsTransformation": "true"} if allow_crs_transformation else None
-        return self._create_multiple(feature, resource_path=resource_path, cls=FeatureList, params=params)
+        extra_body_fields = {"allowCrsTransformation": "true"} if allow_crs_transformation else {}
+        return self._create_multiple(
+            feature, resource_path=resource_path, cls=FeatureList, extra_body_fields=extra_body_fields
+        )
 
     @_with_cognite_domain
     def search_features(
         self,
         feature_type: FeatureType,
         filter: Dict[str, Any],
-        attributes: Dict[str, Any] = None,
+        properties: Dict[str, Any] = None,
         limit: int = 100,
         order_by: List[OrderSpec] = None,
         allow_crs_transformation: bool = False,
@@ -311,7 +320,7 @@ class ExperimentalGeospatialAPI(APIClient):
             feature_type: the feature type to search for
             filter (Dict[str, Any]): the search filter
             limit (int): maximum number of results
-            attributes (Dict[str, Any]): the output attribute selection
+            properties (Dict[str, Any]): the output property selection
             order_by (List[OrderSpec]): the order specification
             allow_crs_transformation: If true, then input geometries will be transformed into the Coordinate Reference
                 System defined in the feature type specification. When it is false, then requests with geometries in
@@ -329,13 +338,13 @@ class ExperimentalGeospatialAPI(APIClient):
                 >>> c = CogniteClient()
                 >>> my_feature_type = c.geospatial.retrieve_feature_types(external_id="my_feature_type")
                 >>> my_feature = c.geospatial.create_features(my_feature_type, Feature(external_id="my_feature", temperature=12.4))
-                >>> res = c.geospatial.search_features(my_feature_type, filter={"range": {"attribute": "temperature", "gt": 12.0}})
+                >>> res = c.geospatial.search_features(my_feature_type, filter={"range": {"property": "temperature", "gt": 12.0}})
                 >>> for f in res:
                 ...     # do something with the features
 
-            Search for features and select output attributes:
+            Search for features and select output properties:
 
-                >>> res = c.geospatial.search_features(my_feature_type, filter={}, attributes={"temperature": {}, "pressure": {}})
+                >>> res = c.geospatial.search_features(my_feature_type, filter={}, properties={"temperature": {}, "pressure": {}})
 
             Search for features and order results:
 
@@ -344,12 +353,16 @@ class ExperimentalGeospatialAPI(APIClient):
         """
         resource_path = self._feature_resource_path(feature_type) + "/search"
         cls = FeatureList
-        order = None if order_by is None else [f"{item.attribute}:{item.direction}" for item in order_by]
-        params = {"allowCrsTransformation": "true"} if allow_crs_transformation else None
+        order = None if order_by is None else [f"{item.property}:{item.direction}" for item in order_by]
         res = self._post(
             url_path=resource_path,
-            json={"filter": filter, "limit": limit, "output": {"attributes": attributes}, "sort": order},
-            params=params,
+            json={
+                "filter": filter,
+                "limit": limit,
+                "output": {"properties": properties},
+                "sort": order,
+                "allowCrsTransformation": (True if allow_crs_transformation else None),
+            },
         )
         return cls._load(res.json()["items"], cognite_client=self._cognite_client)
 
@@ -357,7 +370,7 @@ class ExperimentalGeospatialAPI(APIClient):
         self,
         feature_type: FeatureType,
         filter: Dict[str, Any],
-        attributes: Dict[str, Any] = None,
+        properties: Dict[str, Any] = None,
         allow_crs_transformation: bool = False,
     ) -> Generator[Feature, None, None]:
         """`Stream features`
@@ -366,7 +379,7 @@ class ExperimentalGeospatialAPI(APIClient):
         Args:
             feature_type: the feature type to search for
             filter (Dict[str, Any]): the search filter
-            attributes (Dict[str, Any]): the output attribute selection
+            properties (Dict[str, Any]): the output property selection
             allow_crs_transformation: If true, then input geometries will be transformed into the Coordinate Reference
                 System defined in the feature type specification. When it is false, then requests with geometries in
                 Coordinate Reference System different from the ones defined in the feature type will result in
@@ -383,19 +396,19 @@ class ExperimentalGeospatialAPI(APIClient):
                 >>> c = CogniteClient()
                 >>> my_feature_type = c.geospatial.retrieve_feature_types(external_id="my_feature_type")
                 >>> my_feature = c.geospatial.create_features(my_feature_type, Feature(external_id="my_feature", temperature=12.4))
-                >>> features = c.geospatial.stream_features(my_feature_type, filter={"range": {"attribute": "temperature", "gt": 12.0}})
+                >>> features = c.geospatial.stream_features(my_feature_type, filter={"range": {"property": "temperature", "gt": 12.0}})
                 >>> for f in features:
                 ...     # do something with the features
 
-            Stream features and select output attributes:
+            Stream features and select output properties:
 
-                >>> features = c.geospatial.stream_features(my_feature_type, filter={}, attributes={"temperature": {}, "pressure": {}})
+                >>> features = c.geospatial.stream_features(my_feature_type, filter={}, properties={"temperature": {}, "pressure": {}})
                 >>> for f in features:
                 ...     # do something with the features
 
         """
         resource_path = self._feature_resource_path(feature_type) + "/search-streaming"
-        json = {"filter": filter, "output": {"attributes": attributes, "jsonStreamFormat": "NEW_LINE_DELIMITED"}}
+        json = {"filter": filter, "output": {"properties": properties, "jsonStreamFormat": "NEW_LINE_DELIMITED"}}
         params = {"allowCrsTransformation": "true"} if allow_crs_transformation else None
 
         self._config.headers.pop(self.X_COGNITE_DOMAIN, None)
@@ -406,15 +419,18 @@ class ExperimentalGeospatialAPI(APIClient):
         )
         self._config.headers.pop(self.X_COGNITE_DOMAIN, None)
 
-        for line in res.iter_lines():
-            yield Feature._load(complexjson.loads(line))
+        try:
+            for line in res.iter_lines():
+                yield Feature._load(complexjson.loads(line))
+        except (ChunkedEncodingError, ConnectionError) as e:
+            raise CogniteConnectionError(e)
 
     @_with_cognite_domain
     def aggregate_features(
         self,
         feature_type: FeatureType,
         filter: Dict[str, Any],
-        attribute: str,
+        property: str,
         aggregates: List[str],
         group_by: List[str] = None,
     ) -> FeatureAggregateList:
@@ -424,22 +440,22 @@ class ExperimentalGeospatialAPI(APIClient):
         Args:
             feature_type: the feature type to filter features from
             filter (Dict[str, Any]): the search filter
-            attribute (str): the attribute for which aggregates should be calculated
+            property (str): the property for which aggregates should be calculated
             aggregates (List[str]): list of aggregates to be calculated
-            group_by (List[str]): list of attributes to group by with
+            group_by (List[str]): list of properties to group by with
 
         Returns:
             FeatureAggregateList: the filtered features
 
         Examples:
 
-            Aggregate attribute of features:
+            Aggregate property of features:
 
                 >>> from cognite.experimental import CogniteClient
                 >>> c = CogniteClient()
                 >>> my_feature_type = c.geospatial.retrieve_feature_types(external_id="my_feature_type")
                 >>> my_feature = c.geospatial.create_features(my_feature_type, Feature(external_id="my_feature", temperature=12.4))
-                >>> res = c.geospatial.aggregate_features(my_feature_type, filter={"range": {"attribute": "temperature", "gt": 12.0}}, attribute="temperature", aggregates=["max", "min"], groupBy=["category"])
+                >>> res = c.geospatial.aggregate_features(my_feature_type, filter={"range": {"property": "temperature", "gt": 12.0}}, property="temperature", aggregates=["max", "min"], groupBy=["category"])
                 >>> for a in res:
                 ...     # loop over aggregates in different groups
 
@@ -448,7 +464,7 @@ class ExperimentalGeospatialAPI(APIClient):
         cls = FeatureAggregateList
         res = self._post(
             url_path=resource_path,
-            json={"filter": filter, "attribute": attribute, "aggregates": aggregates, "groupBy": group_by},
+            json={"filter": filter, "property": property, "aggregates": aggregates, "groupBy": group_by},
         )
         return cls._load(res.json()["items"], cognite_client=self._cognite_client)