+++ CHANGELOG.md
@@ -12,6 +12,22 @@ Changes are grouped as follows
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## [0.60.36]
+
+### Changed
+- Geospatial: breaking change, attribute becomes property, following the api terminology
+
+## [0.60.35]
+
+### Fixed
+- Geospatial: log should not consume streaming content.
+- Geospatial: throw CogniteConnectionError when streaming connection closes 
+
+## [0.60.34]
+
+### Fixed
+- Geospatial: follow api changes regarding allowCrsTransformation and delete recursive options.
+
 ## [0.60.33]
 
 ### Changed